[TabsManager] Separate VsCommandFilter logic.
